requirements.txt
amadeus==9.0.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.111.0
uvicorn==0.30.0
aiofiles>=23.0.0
crawl4ai>=0.8.0
mcp>=1.0.0
